rename audit module
